OrderService update method implementation + tests

+++ src/main/java/com/enterpriseintellijence/enterpriseintellijence/data/services/OrderServiceImp.java
@@ -6,21 +6,20 @@ import com.enterpriseintellijence.enterpriseintellijence.data.repository.OrderRe
 import com.enterpriseintellijence.enterpriseintellijence.dto.OrderCreateDTO;
 import com.enterpriseintellijence.enterpriseintellijence.dto.OrderDTO;
 import com.enterpriseintellijence.enterpriseintellijence.dto.UserDTO;
+import com.enterpriseintellijence.enterpriseintellijence.dto.enums.OrderState;
 import com.enterpriseintellijence.enterpriseintellijence.exception.IdMismatchException;
-import com.enterpriseintellijence.enterpriseintellijence.security.JwtContextUtils;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.fge.jsonpatch.JsonPatch;
-import com.github.fge.jsonpatch.JsonPatchException;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 
 import org.modelmapper.ModelMapper;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.time.Clock;
 import java.time.LocalDateTime;
+import java.util.Arrays;
 
 @Service
 @RequiredArgsConstructor
@@ -29,9 +28,7 @@ public class OrderServiceImp implements OrderService {
     private final OrderRepository orderRepository;
     private final UserService userService;
     private final ModelMapper modelMapper;
-
     private final Clock clock;
-
 
     @Override
     public OrderDTO createOrder(OrderCreateDTO orderDTO) {
@@ -49,6 +46,7 @@ public class OrderServiceImp implements OrderService {
     @Override
     public OrderDTO replaceOrder(String id, OrderDTO orderDTO) throws IllegalAccessException {
         throwOnIdMismatch(id, orderDTO);
+
         Order oldOrder = orderRepository.findById(id).orElseThrow(EntityNotFoundException::new);
         Order newOrder = mapToEntity(orderDTO);
 
@@ -58,9 +56,11 @@ public class OrderServiceImp implements OrderService {
         if(!requestingUser.getId().equals(oldOrder.getUser().getId())) {
             throw new IllegalAccessException("User cannot change order");
         }
+
         if(!requestingUser.getId().equals(newOrder.getUser().getId())) {
             throw new IllegalAccessException("User cannot change order");
         }
+
         if(!oldOrder.getState().equals(newOrder.getState())) {
             throw new IllegalAccessException("State cannot be changed");
         }
@@ -70,16 +70,87 @@ public class OrderServiceImp implements OrderService {
     }
 
     @Override
-    public OrderDTO updateOrder(String id, JsonPatch patch) throws JsonPatchException {
-        OrderDTO order = mapToDTO(orderRepository.findById(id).orElseThrow(EntityNotFoundException::new));
-        order = applyPatch(patch, mapToEntity(order));
-        orderRepository.save(mapToEntity(order));
-        return order;
+    public OrderDTO updateOrder(String id, OrderDTO patch) throws IllegalAccessException {
+
+        OrderDTO orderDTO = mapToDTO(orderRepository.findById(id).orElseThrow(EntityNotFoundException::new));
+
+        UserDTO userDTO = userService.findUserFromContext()
+                .orElseThrow(EntityNotFoundException::new);
+
+        if (!userDTO.getId().equals(orderDTO.getUser().getId())) {
+            throw new IllegalAccessException("User cannot change order");
+        }
+
+        if (Arrays.asList(OrderState.values()).contains(patch.getState())) {
+
+            switch (orderDTO.getState()) {
+
+                case CANCELED -> {
+                    if (!patch.getState().equals(OrderState.COMPLETED)) {
+                        throw new IllegalAccessException("State cannot be changed");
+                    }
+                }
+
+                case PENDING -> {
+                    if (!(patch.getState().equals(OrderState.CANCELED) ||
+                        patch.getState().equals(OrderState.PURCHASED))) {
+                        throw new IllegalAccessException("State cannot be changed");
+                    }
+                }
+
+                case PURCHASED -> {
+                    if (patch.getState().equals(OrderState.PENDING) ||
+                        patch.getState().equals(OrderState.DELIVERED) ||
+                        patch.getState().equals(OrderState.COMPLETED) ||
+                        patch.getState().equals(OrderState.CANCELED)) {
+                        throw new IllegalAccessException("State cannot be changed");
+                    }
+                }
+
+                case SHIPPED -> {
+                    if (patch.getState().equals(OrderState.PENDING) ||
+                        patch.getState().equals(OrderState.PURCHASED) ||
+                        patch.getState().equals(OrderState.COMPLETED) ||
+                        patch.getState().equals(OrderState.CANCELED)) {
+                        throw new IllegalAccessException("State cannot be changed");
+                    }
+                }
+
+                case DELIVERED -> {
+                    if (patch.getState().equals(OrderState.PENDING) ||
+                        patch.getState().equals(OrderState.PURCHASED) ||
+                        patch.getState().equals(OrderState.SHIPPED) ||
+                        patch.getState().equals(OrderState.CANCELED)) {
+                        throw new IllegalAccessException("State cannot be changed");
+                    }
+                }
+
+                case COMPLETED -> {
+                    if (!patch.getState().equals(OrderState.CANCELED)) {
+                        throw new IllegalAccessException("State cannot be changed");
+                    }
+                }
+            }
+
+            orderDTO.setState(patch.getState());
+
+        } else {
+            throw new IllegalAccessException("State cannot be changed");
+        }
+
+        orderDTO = mapToDTO(orderRepository.save(mapToEntity(orderDTO)));
+        return orderDTO;
     }
 
     @Override
-    public OrderDTO deleteOrder(String id) {
+    public OrderDTO deleteOrder(String id) throws IllegalAccessException {
         Order order = orderRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+        UserDTO user = userService.findUserFromContext()
+            .orElseThrow(EntityNotFoundException::new);
+
+        if (!user.getId().equals(order.getUser().getId())) {
+            throw new IllegalAccessException("User cannot delete order");
+        }
 
         orderRepository.deleteById(id);
         return mapToDTO(order);
@@ -97,16 +168,17 @@ public class OrderServiceImp implements OrderService {
         return orderDTO;
     }
 
+    public Page<OrderDTO> findAllByUserId(Pageable pageable) {
+
+        UserDTO userDTO = userService.findUserFromContext()
+            .orElseThrow(EntityNotFoundException::new);
+
+        return orderRepository.findAllByUserId(userDTO.getId(), pageable).map(this::mapToDTO);
+    }
+
     private void throwOnIdMismatch(String id, OrderDTO orderDTO) {
         if (orderDTO.getId() != null && !orderDTO.getId().equals(id))
             throw new IdMismatchException();
-    }
-
-    public OrderDTO applyPatch(JsonPatch patch, Order order) throws JsonPatchException {
-        ObjectMapper objectMapper = new ObjectMapper();
-        JsonNode patched = patch.apply(objectMapper.convertValue(order, JsonNode.class));
-
-        return objectMapper.convertValue(patched, OrderDTO.class);
     }
 
     public Order mapToEntity(OrderDTO orderDTO) {
@@ -120,6 +192,4 @@ public class OrderServiceImp implements OrderService {
     public OrderDTO mapToDTO(Order order) {
         return modelMapper.map(order, OrderDTO.class);
     }
-
-
 }
